+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
 - Upgraded ska-mid-ds-manager version to v2.0.0
 - Added actStaticOffsetValueXel and actStaticOffsetValueEl attributes
 - Updated band<N>PointingModelParams usage
+- Added `lastCommandedMode` attribute to record the last mode change request
 
 Version 3.0.1
 *************